+++ app/main.py
@@ -7,7 +7,7 @@ app = FastAPI(title="Asistencia API")
 @app.on_event("startup")
 def startup_event():
     init_db()
-    test_connection()
+    #test_connection()
 
 # Registrar rutas
 app.include_router(asistencia.router, prefix="/asistencia", tags=["Asistencia"])